Added custom audio device example
 Added custom audio device example

Added custom audio device example

Works

Updated to kotlin(wip)

compiled

Added README

Updated images

ci pipeline changes. changed the ui changes when a participant already exists

Reverted unwanted changes

Reverted unwanted changes

Updated README

reverted unwanted chnages

refactor package name
