+++ build.gradle
@@ -1,5 +1,5 @@
 buildscript {
-    ext.kotlin_version = '1.3.61'
+    ext.kotlin_version = '1.4.0'
     ext.versions = [
             'java'               : JavaVersion.VERSION_1_8,
             'androidGradlePlugin': '4.0.1',
@@ -14,7 +14,7 @@ buildscript {
             'retrofit'           : '2.0.0-beta4',
             'okhttp'             : '3.6.0',
             'ion'                : '2.1.8',
-            'videoAndroid'       : '5.11.1',
+            'videoAndroid'       : '5.12.0',
             'audioSwitch'        : '1.0.1'
     ]
