--- exampleAudioSink/src/main/java/com/twilio/exampleaudiosink/dialog/Dialog.java
@@ -1,35 +1,0 @@
-package com.twilio.exampleaudiosink.dialog;
-
-import android.content.Context;
-import android.content.DialogInterface;
-import android.support.v7.app.AlertDialog;
-import android.widget.EditText;
-
-import com.twilio.exampleaudiosink.R;
-
-public class Dialog {
-
-    public static AlertDialog createConnectDialog(EditText participantEditText,
-                                                  DialogInterface.OnClickListener callParticipantsClickListener,
-                                                  DialogInterface.OnClickListener cancelClickListener,
-                                                  Context context) {
-        AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(context);
-
-        alertDialogBuilder.setIcon(android.R.drawable.sym_call_outgoing);
-        alertDialogBuilder.setTitle("Connect to a room");
-        alertDialogBuilder.setPositiveButton("Connect", callParticipantsClickListener);
-        alertDialogBuilder.setNegativeButton("Cancel", cancelClickListener);
-        alertDialogBuilder.setCancelable(false);
-
-        setRoomNameFieldInDialog(participantEditText, alertDialogBuilder);
-
-        return alertDialogBuilder.create();
-    }
-
-    private static void setRoomNameFieldInDialog(EditText roomNameEditText,
-                                                 AlertDialog.Builder alertDialogBuilder) {
-        roomNameEditText.setHint("room name");
-        alertDialogBuilder.setView(roomNameEditText);
-    }
-
-}
