+++ exampleAudioSink/src/main/java/com/twilio/examplecustomaudiodevice/dialog/Dialog.java
@@ -1,0 +1,33 @@
+package com.twilio.examplecustomaudiodevice.dialog;
+
+import android.content.Context;
+import android.content.DialogInterface;
+import android.support.v7.app.AlertDialog;
+import android.widget.EditText;
+
+public class Dialog {
+
+    public static AlertDialog createConnectDialog(EditText participantEditText,
+                                                  DialogInterface.OnClickListener callParticipantsClickListener,
+                                                  DialogInterface.OnClickListener cancelClickListener,
+                                                  Context context) {
+        AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(context);
+
+        alertDialogBuilder.setIcon(android.R.drawable.sym_call_outgoing);
+        alertDialogBuilder.setTitle("Connect to a room");
+        alertDialogBuilder.setPositiveButton("Connect", callParticipantsClickListener);
+        alertDialogBuilder.setNegativeButton("Cancel", cancelClickListener);
+        alertDialogBuilder.setCancelable(false);
+
+        setRoomNameFieldInDialog(participantEditText, alertDialogBuilder);
+
+        return alertDialogBuilder.create();
+    }
+
+    private static void setRoomNameFieldInDialog(EditText roomNameEditText,
+                                                 AlertDialog.Builder alertDialogBuilder) {
+        roomNameEditText.setHint("room name");
+        alertDialogBuilder.setView(roomNameEditText);
+    }
+
+}
