--- exampleAudioSink/src/test/java/com/twilio/exampleaudiosink/ExampleUnitTest.java
@@ -1,17 +1,0 @@
-package com.twilio.exampleaudiosink;
-
-import org.junit.Test;
-
-import static org.junit.Assert.assertEquals;
-
-/**
- * Example local unit test, which will execute on the development machine (host).
- *
- * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
- */
-public class ExampleUnitTest {
-    @Test
-    public void addition_isCorrect() {
-        assertEquals(4, 2 + 2);
-    }
-}
