+++ exampleCustomAudioDevice/proguard-rules.pro
@@ -1,0 +1,4 @@
+-keep class tvi.webrtc.** { *; }
+-keep class com.twilio.video.** { *; }
+-keep class com.twilio.common.** { *; }
+-keepattributes InnerClasses
