+++ exampleCustomAudioDevice/src/main/res/values/dimens.xml
@@ -1,0 +1,4 @@
+<resources>
+    <!-- Default screen margins, per the Android Design guidelines. -->
+    <dimen name="fab_margin">16dp</dimen>
+</resources>
