+++ exampleCustomAudioDevice/src/main/res/values/strings.xml
@@ -1,0 +1,12 @@
+<resources>
+    <string name="app_name">exampleCustomAudioDevice</string>
+    <string name="error_retrieving_access_token">Error retrieving token</string>
+    <string name="permissions_needed">Microphone permission needed. Please allow at App start for additional functionality.</string>
+
+    <!-- Strings related to Settings -->
+
+    <string name="status_two_particpants_needed">Please join on another device. This example only works with two participants</string>
+    <string name="status_join_room">Please join a room with another participant</string>
+    <string name="status_capture_ready">RemotePartipant joined and will hear music by default. Press the play/pause button to toggle between microphone and music file.</string>
+    <string name="connect_failed">Failed to connect</string>
+</resources>
